+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Show preview class file versions in the structure view
 - Show annotation method default values in the structure view
 - Show inner class details in the structure view
+- Show try-catch type annotations in the structure view
 
 ### Changed