+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Fixed
 
+- Fix reload of the structure view didn't preserve the expanded tree nodes
 - Fix invalid tool window data values breaking action updates
 - Fix opening the structure view's class file in the editor
 - Fix toolbar creation triggering IntelliJ Platform action-group errors