+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Fix "Analyze Current Editor File" silently doing nothing when no processable editor file is open
 - Fix "Analyze Byte Code" processing only one item from some project-view multi-selections
 - Fix open file actions will now open the file explorer and not the browser
+- Fix access converter checkboxes showing as text
 
 ## 4.4.0 - 2025-12-04